+++ copybook-governance-eval/Cargo.toml
@@ -1,6 +1,6 @@
 # SPDX-License-Identifier: AGPL-3.0-or-later
 [package]
-name = "copybook-governance-runtime"
+name = "copybook-governance-eval"
 version.workspace = true
 edition.workspace = true
 rust-version.workspace = true
@@ -8,15 +8,16 @@ license.workspace = true
 authors.workspace = true
 repository.workspace = true
 homepage.workspace = true
-description = "Runtime governance state evaluation for support-matrix and feature-flag interoperability."
-documentation = "https://docs.rs/copybook-governance-runtime"
+description = "Runtime governance state evaluation logic extracted for SRP-focused reuse."
+documentation = "https://docs.rs/copybook-governance-eval"
 readme = "README.md"
 include = ["src/**", "Cargo.toml", "README.md", "LICENSE"]
 keywords = ["cobol", "copybook", "feature-flags", "support-matrix", "runtime"]
 categories = ["development-tools", "data-management"]
 
 [dependencies]
-copybook-governance-eval.workspace = true
+copybook-governance-grid.workspace = true
+serde.workspace = true
 
 [lints]
 workspace = true
