+++ copybook-governance-eval/README.md
@@ -1,0 +1,9 @@
+# copybook-governance-eval
+
+SRP microcrate for runtime governance evaluation:
+
+- Resolves support-matrix entries against active feature flags
+- Produces interoperable governance state rows and runtime summaries
+- Reuses canonical governance-grid feature and support contracts
+
+`copybook-governance-runtime` remains as the compatibility/re-export layer.
